+++ tests/jobname.py
@@ -7,7 +7,7 @@ from pylatex import Document
 
 
 def test():
-    doc = Document('jobname_test', title='Jobname test', maketitle=True)
+    doc = Document('jobname_test', data=['Jobname test'])
     doc.generate_pdf()
 
     assert os.path.isfile('jobname_test.pdf')
@@ -18,7 +18,7 @@ def test():
     os.makedirs(folder)
     path = os.path.join(folder, 'jobname_test_dir')
 
-    doc = Document(path, title='Jobname test dir', maketitle=True)
+    doc = Document(path, data=['Jobname test dir'])
     doc.generate_pdf()
 
     assert os.path.isfile(path + '.pdf')
@@ -28,7 +28,7 @@ def test():
     os.makedirs(folder)
     path = os.path.join(folder, 'jobname_test_dir2')
 
-    doc = Document(path, title='Jobname test dir', maketitle=True)
+    doc = Document(path, data=['Jobname test dir'])
     doc.generate_pdf(os.path.join(folder, ''))
 
     assert os.path.isfile(path + '.pdf')